cds2kaks

This function calculates Ka/Ks (dN/dS; accoring to Li (1993) or Yang and Nielson (2000) or Nei and Gojobori (1986) given two coding sequences cds1 and cds2 as DNAStringSet or DNAString.

cds2kaks(
  cds1,
  cds2,
  model = "Li",
  kakscalcpath = paste0(find.package("CRBHits"), "/extdata/KaKs_Calculator2.0_src/src/"),
  ...
)

Arguments

cds1

single cds1 sequence as DNAStringSet or DNAString [mandatory]

cds2

single cds2 sequence as DNAStringSet or DNAString [mandatory]

model

specify codon model either "Li" or "NG86" or one of KaKs_Calculator2 model "NG", "LWL", "LPB", "MLWL", "MLPB", "GY", "YN", "MYN", "MS", "MA", "GNG", "GLWL", "GLPB", "GMLWL", "GMLPB", "GYN", "GMYN" [default: Li]

kakscalcpath

specify the PATH to the KaKs_Calculator binaries [default: /extdata/KaKs_Calculator2.0_src/src/]

...

other codon alignment parameters (see cds2codonaln)

Value

vector of ka/ks values as specified by seqinr::kaks or KaKs_Calculator2.0

References

Li WH. (1993) Unbiased estimation of the rates of synonymous and nonsynonymous substitution. J. Mol. Evol., 36, 96-99.

Wang D, Zhang Y et al. (2010) KaKs_Calculator 2.0: a toolkit incorporating gamma-series methods and sliding window strategies. Genomics Proteomics Bioinformatics. 8(1), 77-80.

Yang Z and Nielson R. (2000) Estimating synonymous and nonsynonymous substitution rates under realistic evolutionary models. Mol. Biol. Evol., 17(1), 32-43.

Nei and Gojobori. (1986) Simple methods for estimating the numbers of synonymous and nonsynonymous nucleotide substitutions. Mol. Biol. Evol., 3(5), 418-426.

See also

Author

Kristian K Ullrich

Examples

## load example sequence data
data("ath", package="CRBHits")
data("aly", package="CRBHits")
## select a sequence pair according to a best hit pair (done for you)
cds1 <- ath[1]
cds2 <- aly[282]
cds2kaks(
    cds1=cds1,
    cds2=cds2,
    model="Li")
#>         ka         ks        vka        vks 
#> 1.02338181 9.99999900 0.01140267 9.99999900 
## using an alternative substitutionMatrix
cds2kaks(
    cds1=cds1,
    cds2=cds2,
    model="Li",
    substitutionMatrix="BLOSUM45")
#>          ka          ks         vka         vks 
#> 0.878034285 9.999999000 0.009456807 9.999999000 
cds2kaks(
    cds1=cds1,
    cds2=cds2,
    model="Li",
    substitutionMatrix="BLOSUM80")
#>          ka          ks         vka         vks 
#> 0.814370837 9.999999000 0.009489614 9.999999000 
if (FALSE) {
cds2kaks(
    cds1=cds1,
    cds2=cds2,
    model="YN")
}
if (FALSE) {
cds2kaks(
    cds1=cds1,
    cds2=cds2,
    model="NG86")
}