cdsfile2rbh

This function calculates (conditional-)reciprocal best hit pair matrix from two two cds fasta file's.

cdsfile2rbh(
  cdsfile1,
  cdsfile2,
  lastpath = paste0(find.package("CRBHits"), "/extdata/last-1060/src/"),
  outpath = "/tmp",
  crbh = TRUE,
  keepSingleDirection = FALSE,
  evalue = 0.001,
  qcov = 0,
  tcov = 0,
  pident = 0,
  alnlen = 0,
  rost1999 = FALSE,
  filter = NULL,
  plotCurve = FALSE,
  threads = 1,
  remove = TRUE
)

Arguments

cdsfile1
cds1 fasta file [mandatory]

cdsfile2
cds2 fasta file [mandatory]

lastpath
specify the PATH to the last binaries [default: /extdata/last-1060/src/]

outpath
specify the output PATH [default: /tmp]

crbh
specify if conditional-reciprocal hit pairs should be retained as secondary hits [default: TRUE]

keepSingleDirection
specify if single direction secondary hit pairs should be retained [default: FALSE]

evalue
evalue [default: 1e-3]

qcov
query coverage [default: 0.0]

tcov
target coverage [default: 0.0]

pident
percent identity [default: 0.0]

alnlen
alignment length [default: 0.0]

rost1999
specifiy if hit pairs should be filter by equation 2 of Rost 1999 [default: FALSE]

filter
specify additonal custom filters to be applied on hit pairs [default: NULL]

plotCurve
specify if crbh fitting curve should be plotted [default: FALSE]

threads
number of parallel threads [default: 1]

remove
specify if last result files should be removed [default: TRUE]

Value

List of three 1: $crbh.pairs (crbh = TRUE) or $rbh.pairs (crbh = FALSE) 2: $crbh1 matrix (crbh = TRUE) or $rbh1 (crbh = FALSE); query > target 3: $crbh2 matrix (crbh = TRUE) or $rbh2 (crbh = FALSE); target > query

References

Aubry S, Kelly S et al. (2014) Deep Evolutionary Comparison of Gene Expression Identifies Parallel Recruitment of Trans-Factors in Two Independent Origins of C4 Photosynthesis. PLOS Genetics, 10(6) e1004365.

Kiełbasa, SM et al. (2011) Adaptive seeds tame genomic sequence comparison. Genome research, 21(3), 487-493.

Rost B. (1999). Twilight zone of protein sequence alignments. Protein Engineering, 12(2), 85-94.

Examples

##load example sequence data
athfile <- paste0(find.package("CRBHits"), "inst/fasta/ath.cds.fasta.gz")
alyfile <- paste0(find.package("CRBHits"), "inst/fasta/aly.cds.fasta.gz")
#conditional-reciprocal best hits
ath_aly_crbh <- cdsfile2rbh(athfile, alyfile, plotCurve = TRUE)

#> Error in cdsfile2rbh(athfile, alyfile, plotCurve = TRUE): Error: last PATH does not exist. Please specify correct PATH and/or look into package installation prerequisites. Try to use make.last() function.

dim(ath_aly_crbh$crbh.pairs)

#> [1] 798   2

#reciprocal best hits
ath_aly_rbh <- cdsfile2rbh(athfile, alyfile, crbh = FALSE)

#> Error in cdsfile2rbh(athfile, alyfile, crbh = FALSE): Error: last PATH does not exist. Please specify correct PATH and/or look into package installation prerequisites. Try to use make.last() function.

dim(ath_aly_rbh$rbh.pairs)

#> Error in eval(expr, envir, enclos): object 'ath_aly_rbh' not found