dnastring2kaks

This function calculates Ka/Ks (pN/pS; according to Li (1993) or Nei and Gojobori (1986) for all combinations of a DNAStringSet.

dnastring2kaks(cds, align = FALSE, model = "Li", threads = 1, ...)

Arguments

cds
DNAStringSet A [mandatory]

align
specify if sequence combinations should be aligned pairwise by cds2codonaln. Default assumes sequences are aligned [default: FALSE]

model
specify codon model either "Li" or "NG86" [default: Li]

threads
number of parallel threads [default: 1]

...
other codon alignment parameters (see cds2codonaln)

Value

A data.frame of KaKs values

References

Nei and Gojobori. (1986) Simple methods for estimating the numbers of synonymous and nonsynonymous nucleotide substitutions. Mol. Biol. Evol., 3(5), 418-426.

Ganeshan et al. (1997) Human immunodeficiency virus type 1 genetic evolution in children with different rates of development of disease. J. Virology. 71(1), 663-677.

Yang et al. (2000) Codon-substitution models for heterogeneous selection pressure at amino acid sites. Genetics. 155(1), 431-449.

See also

Author

Kristian K Ullrich

Examples

## load example sequence data
data("hiv", package="CRBHits")
dnastring2kaks(hiv, model = "Li")

#> Joining, by = c("Comp1", "Comp2")

#> Joining, by = c("Comp1", "Comp2")

#> Joining, by = c("Comp1", "Comp2")

#>     Comp1  Comp2         ka            ks          vka          vks
#> 1  U68496 U68497 0.03026357  0.0317031856 0.0003051202 0.0004007730
#> 2  U68496 U68498 0.09777332  0.0176141589 0.0009314173 0.0005091970
#> 3  U68496 U68499 0.10295875  0.0176731064 0.0009595527 0.0005787387
#> 4  U68496 U68500 0.13461355  0.0463968986 0.0013731885 0.0020497481
#> 5  U68496 U68501 0.12607831  0.0284429395 0.0013277282 0.0006310195
#> 6  U68496 U68502 0.17441037  0.1092653245 0.0017871934 0.0040766687
#> 7  U68496 U68503 0.15634972  0.0397478904 0.0015693139 0.0010151111
#> 8  U68496 U68504 0.14421633  0.0879405817 0.0014555647 0.0036120005
#> 9  U68496 U68505 0.13952943  0.0600004376 0.0013871757 0.0010872124
#> 10 U68496 U68506 0.15293806  0.0446263293 0.0014917205 0.0016000455
#> 11 U68496 U68507 0.17233755  0.0506739094 0.0017489395 0.0008525837
#> 12 U68496 U68508 0.20297426  0.0530563367 0.0021784791 0.0016191407
#> 13 U68497 U68498 0.12291500  0.0178006922 0.0011451050 0.0005798517
#> 14 U68497 U68499 0.11423944  0.0178657773 0.0010645577 0.0006536140
#> 15 U68497 U68500 0.14635158  0.0465803514 0.0014882273 0.0021421943
#> 16 U68497 U68501 0.15232262  0.0289560070 0.0015646067 0.0007113171
#> 17 U68497 U68502 0.18844835  0.1082484036 0.0019299056 0.0043143765
#> 18 U68497 U68503 0.15389053  0.0398280135 0.0015431390 0.0011502503
#> 19 U68497 U68504 0.15616528  0.1109611491 0.0015766080 0.0039369110
#> 20 U68497 U68505 0.15136945  0.0606006545 0.0015059746 0.0011657538
#> 21 U68497 U68506 0.16502115  0.0782677755 0.0016165774 0.0020628254
#> 22 U68497 U68507 0.20090792  0.0518729513 0.0020496504 0.0009519639
#> 23 U68497 U68508 0.20011790  0.0534260717 0.0021725951 0.0017341869
#> 24 U68498 U68499 0.07596391 -0.0004838245 0.0005620535 0.0003865916
#> 25 U68498 U68500 0.08840278  0.0463667477 0.0008260570 0.0016460465
#> 26 U68498 U68501 0.10286217  0.0101958272 0.0010525516 0.0003166930
#> 27 U68498 U68502 0.13009195  0.0940491643 0.0013057579 0.0044558026
#> 28 U68498 U68503 0.13641935  0.0441729344 0.0013004754 0.0016970196
#> 29 U68498 U68504 0.11024181  0.1184319156 0.0010976172 0.0049864168
#> 30 U68498 U68505 0.10652695  0.0712975708 0.0010391852 0.0023973803
#> 31 U68498 U68506 0.13085907  0.0870858286 0.0013133008 0.0029231671
#> 32 U68498 U68507 0.10809058  0.0543870183 0.0011142558 0.0015017030
#> 33 U68498 U68508 0.16297575  0.0440517806 0.0018854712 0.0010247788
#> 34 U68499 U68500 0.04753693  0.0102737435 0.0004383279 0.0005298658
#> 35 U68499 U68501 0.13602580  0.0097162829 0.0012647818 0.0007716744
#> 36 U68499 U68502 0.13681777  0.0901640402 0.0012227320 0.0042540042
#> 37 U68499 U68503 0.15100361  0.0458293630 0.0012941257 0.0020242775
#> 38 U68499 U68504 0.10903541  0.0941359554 0.0010512661 0.0039210531
#> 39 U68499 U68505 0.11155268  0.0408401022 0.0010437110 0.0009288677
#> 40 U68499 U68506 0.12931635  0.0641003604 0.0012583150 0.0020419709
#> 41 U68499 U68507 0.12466831  0.0567007476 0.0012199686 0.0017900222
#> 42 U68499 U68508 0.13610962  0.0898484362 0.0014962567 0.0041187582
#> 43 U68500 U68501 0.14964410  0.0582042332 0.0015776531 0.0020548121
#> 44 U68500 U68502 0.15988006  0.1438472932 0.0016078164 0.0064428544
#> 45 U68500 U68503 0.15673725  0.0861977448 0.0013892022 0.0037479479
#> 46 U68500 U68504 0.12800773  0.1677553950 0.0013867770 0.0080191008
#> 47 U68500 U68505 0.13063843  0.1085890328 0.0013768721 0.0039717993
#> 48 U68500 U68506 0.14653147  0.1236919447 0.0016022322 0.0051547115
#> 49 U68500 U68507 0.14494020  0.0963918335 0.0014435239 0.0034569489
#> 50 U68500 U68508 0.13591211  0.0806060567 0.0014251451 0.0030925806
#> 51 U68501 U68502 0.14933484  0.0672748223 0.0014052841 0.0034772435
#> 52 U68501 U68503 0.15573005  0.0554867582 0.0013851157 0.0021116389
#> 53 U68501 U68504 0.11320207  0.1420723073 0.0010409458 0.0060663691
#> 54 U68501 U68505 0.12691557  0.0670868311 0.0011168607 0.0020037645
#> 55 U68501 U68506 0.11939683  0.1092378646 0.0011341777 0.0038233059
#> 56 U68501 U68507 0.14111743  0.0661867231 0.0013055455 0.0018710869
#> 57 U68501 U68508 0.15062997  0.0674985077 0.0017283433 0.0012918646
#> 58 U68502 U68503 0.14127514  0.1679513190 0.0012163932 0.0074519537
#> 59 U68502 U68504 0.10809248  0.1522765380 0.0010257402 0.0058858088
#> 60 U68502 U68505 0.10525681  0.1405815063 0.0009871737 0.0056509830
#> 61 U68502 U68506 0.12796255  0.1874086342 0.0012356488 0.0084817971
#> 62 U68502 U68507 0.11030071  0.1651909774 0.0010889044 0.0071876798
#> 63 U68502 U68508 0.13517843  0.1841705167 0.0015314707 0.0079501818
#> 64 U68503 U68504 0.13509941  0.1269048145 0.0011113531 0.0047146155
#> 65 U68503 U68505 0.14189994  0.0424813552 0.0011379089 0.0009375791
#> 66 U68503 U68506 0.15642763  0.0962526692 0.0013415432 0.0027808415
#> 67 U68503 U68507 0.14632577  0.0425317911 0.0011497637 0.0008663924
#> 68 U68503 U68508 0.14702536  0.0786270338 0.0013787524 0.0025009601
#> 69 U68504 U68505 0.02018431  0.0935703588 0.0001353697 0.0042476907
#> 70 U68504 U68506 0.01963868  0.0616993490 0.0002675568 0.0019647032
#> 71 U68504 U68507 0.04645002  0.1121475416 0.0004670412 0.0041104637
#> 72 U68504 U68508 0.10262621  0.1452111714 0.0013299190 0.0066748916
#> 73 U68505 U68506 0.04029444  0.0848476525 0.0004104632 0.0024364085
#> 74 U68505 U68507 0.04627943  0.0203211276 0.0004322275 0.0003654116
#> 75 U68505 U68508 0.10915633  0.0474209893 0.0013291206 0.0021449341
#> 76 U68506 U68507 0.04958974  0.0826465317 0.0004228026 0.0022519366
#> 77 U68506 U68508 0.12249925  0.1139798033 0.0015342708 0.0043738539
#> 78 U68507 U68508 0.11267851  0.0468580415 0.0012934094 0.0019973539

dnastring2kaks(hiv, model = "NG86")

#> Error in dnastring2codonmat(cds): Error: frame needs to be 1 or 2 or 3

if (FALSE) {
dnastring2kaks(hiv, model = "NG86", threads = 2)
}
if (FALSE) {
dnastring2kaks(hiv, model = "NG86", align = TRUE, threads = 2)
dnastring2kaks(hiv, model = "Li", align = TRUE, substitutionMatrix = "BLOSUM45")
}