rbh2kaks

This function calculates ka/ks (dN/dS; accoring to Li (1993) or Yang and Nielson (2000) for each reciprocal best hit pair. The names of the rbh columns must match the names of the corresponding cds1 and cds2 DNAStringSet vectors.

rbh2kaks(
  rbhpairs,
  cds1,
  cds2,
  model = "Li",
  threads = 1,
  kakscalcpath = paste0(find.package("CRBHits"), "/extdata/KaKs_Calculator2.0/src/"),
  ...
)

Arguments

rbhpairs
(conditional-)recirpocal best hit pair matrix [mandatory]

cds1
cds1 sequences as DNAStringSet or url for first crbh pairs column [mandatory]

cds2
cds2 sequences as DNAStringSet or url for second crbh pairs column [mandatory]

model
specify codon model either "Li" or "YN" [default: Li]

threads
number of parallel threads [default: 1]

kakscalcpath
specify the PATH to the KaKs_Calculator binaries [default: /extdata/KaKs_Calculator2.0/src/]

...
other codon alignment parameters (see cds2codonaln)

References

Li WH. (1993) Unbiased estimation of the rates of synonymous and nonsynonymous substitution. J. Mol. Evol., 36, 96-99.

Wang D, Zhang Y et al. (2010) KaKs_Calculator 2.0: a toolkit incorporating gamma-series methods and sliding window strategies. Genomics Proteomics Bioinformatics. 8(1), 77-80.

Yang Z and Nielson R. (2000) Estimating synonymous and nonsynonymous substitution rates under realistic evolutionary models. Mol. Biol. Evol., 17(1), 32-43.

See also

Examples

##load example sequence data
data("ath", package="CRBHits")
data("aly", package="CRBHits")
##load example conditional-reciprocal best hit pair results
data("ath_aly_crbh", package="CRBHits")
ath_aly_crbh.kaks <- rbh2kaks(ath_aly_crbh$crbh.pairs[1:20, ],
                              ath, aly, model = "Li")
head(ath_aly_crbh.kaks)

#>           aa1          aa2         ka        ks          vka          vks
#> 1 AT1G01010.1 AL1G11530.t1 0.10617227 0.2628911 2.637032e-04 0.0023637311
#> 2 AT1G01020.5 AL1G11510.t1 0.05027159 0.1076133 2.572181e-04 0.0017464252
#> 3 AT1G01030.1 AL1G11500.t1 0.05606632 0.2329670 1.808118e-04 0.0026270370
#> 4 AT1G01040.1 AL1G11480.t1 0.01388844 0.1121138 6.907144e-06 0.0001460956
#> 5 AT1G01050.1 AL1G11470.t1 0.00000000 0.1831839 0.000000e+00 0.0030223075
#> 6 AT1G01060.6 AL1G11460.t1 0.09054187 0.1854908 1.323098e-04 0.0007490874