rbh2kaks.YN

This function calculates Ka and Ks (dN and dS; accoring to Yang and Nielson 2000) for each reciprocal best hit pair. The names of the rbh columns must match the names of the corresponding cds1 and cds2 DNAStringSet vectors.

rbh2kaks.YN(
  rbh,
  cds1,
  cds2,
  kakscalcpath = paste0(find.package("CRBHits"), "/extdata/KaKs_Calculator2.0/src/"),
  threads = 1
)

Arguments

rbh
conditional-recirpocal best hit pair matrix [mandatory]

cds1
cds1 sequences as DNAStringSet for first rbh column [mandatory]

cds2
cds2 sequences as DNAStringSet for second rbh column [mandatory]

kakscalcpath
specify the PATH to the KaKs_Calculator binaries [default: /extdata/KaKs_Calculator2.0/src/]

threads
number of parallel threads [default: 1]

References

Wang D, Zhang Y et al. (2010) KaKs_Calculator 2.0: a toolkit incorporating gamma-series methods and sliding window strategies. Genomics Proteomics Bioinformatics. 8(1), 77-80.

Yang Z and Nielson R. (2000) Estimating synonymous and nonsynonymous substitution rates under realistic evolutionary models. Mol. Biol. Evol., 17(1), 32-43.

Examples

##load example sequence data
data("ath", package="CRBHits")
data("aly", package="CRBHits")
##load example conditional-reciprocal best hit pair results
data("ath_aly_crbh", package="CRBHits")
ath_aly_crbh.kaks.YN <- rbh2kaks.YN(ath_aly_crbh$crbh.pairs, ath, aly)

#> Error in rbh2kaks.YN(ath_aly_crbh$crbh.pairs, ath, aly): Error: KaKs_Calculator2.0 PATH does not exist. Please specify correct PATH and/or look into package installation prerequisites. Try to use make.KaKs_Calculator2() function.

head(ath_aly_crbh.kaks.YN)

#> Error in head(ath_aly_crbh.kaks.YN): object 'ath_aly_crbh.kaks.YN' not found