cdsfile2rbh

This function calculates (conditional-)reciprocal best hit pair matrix from two cds fasta file's.

cdsfile2rbh(
  cdsfile1,
  cdsfile2,
  lastpath = paste0(find.package("CRBHits"), "/extdata/last-1060/src/"),
  outpath = "/tmp",
  filter = NULL,
  crbh = TRUE,
  keepSingleDirection = FALSE,
  plotCurve = FALSE,
  threads = 1,
  remove = TRUE
)

Arguments

cdsfile1
cds1 fasta file [mandatory]

cdsfile2
cds2 fasta file [mandatory]

lastpath
specify the PATH to the last binaries [default: /extdata/last-1060/src/]

outpath
specify the output PATH [default: /tmp]

filter
specify the filters to be applied on hit pairs e.g.: c("filter.eval", "filter.rost1999") [default: NULL]

crbh
specify if conditional-reciprocal hit pairs should be retained as secondary hits [default: TRUE]

keepSingleDirection
specify if single direction secondary hit pairs should be retained [default: FALSE]

plotCurve
specify if crbh fitting curve should be plotted [default: FALSE]

threads
number of parallel threads [default: 1]

remove
specify if last result files should be removed [default: TRUE]

References

Aubry S, Kelly S et al. (2014) Deep Evolutionary Comparison of Gene Expression Identifies Parallel Recruitment of Trans-Factors in Two Independent Origins of C4 Photosynthesis. PLOS Genetics, 10, 6 e1004365.

Rost B. (1999). Twilight zone of protein sequence alignments. Protein Engineering, 12(2), 85-94.

Examples

##load example sequence data
data("ath", package="CRBHits")
data("aly", package="CRBHits")
#change default filter param
formals(filter.eval) <- alist(x= , eval=1e-10)