dnastring2kaks

This function calculates Ka/Ks (pN/pS; according to Li (1993) or Nei and Gojobori (1986) for all combinations of a DNAStringSet.

dnastring2kaks(cds, align = FALSE, model = "Li", threads = 1, ...)

Arguments

cds
DNAStringSet A [mandatory]

align
specify if sequence combinations should be aligned pairwise by cds2codonaln. Default assumes sequences are aligned [default: FALSE]

model
specify codon model either "Li" or "NG86" [default: Li]

threads
number of parallel threads [default: 1]

...
other codon alignment parameters (see cds2codonaln)

Value

A data.frame of KaKs values

References

Nei and Gojobori. (1986) Simple methods for estimating the numbers of synonymous and nonsynonymous nucleotide substitutions. Mol. Biol. Evol., 3(5), 418-426.

Ganeshan et al. (1997) Human immunodeficiency virus type 1 genetic evolution in children with different rates of development of disease. J. Virology. 71(1), 663-677.

Yang et al. (2000) Codon-substitution models for heterogeneous selection pressure at amino acid sites. Genetics. 155(1), 431-449.

See also

Author

Kristian K Ullrich

Examples

## load example sequence data
data("hiv", package="CRBHits")
dnastring2kaks(hiv, model = "Li")

#> Joining, by = c("seq1", "seq2")

#> Joining, by = c("seq1", "seq2")

#> Joining, by = c("seq1", "seq2")

#>      seq1   seq2         ka            ks          vka          vks
#> 1  U68496 U68497 0.03026357  0.0317031856 0.0003051202 0.0004007730
#> 2  U68496 U68498 0.09777332  0.0176141589 0.0009314173 0.0005091970
#> 3  U68496 U68499 0.10295875  0.0176731064 0.0009595527 0.0005787387
#> 4  U68496 U68500 0.13461355  0.0463968986 0.0013731885 0.0020497481
#> 5  U68496 U68501 0.12607831  0.0284429395 0.0013277282 0.0006310195
#> 6  U68496 U68502 0.17441037  0.1092653245 0.0017871934 0.0040766687
#> 7  U68496 U68503 0.15634972  0.0397478904 0.0015693139 0.0010151111
#> 8  U68496 U68504 0.14421633  0.0879405817 0.0014555647 0.0036120005
#> 9  U68496 U68505 0.13952943  0.0600004376 0.0013871757 0.0010872124
#> 10 U68496 U68506 0.15293806  0.0446263293 0.0014917205 0.0016000455
#> 11 U68496 U68507 0.17233755  0.0506739094 0.0017489395 0.0008525837
#> 12 U68496 U68508 0.20297426  0.0530563367 0.0021784791 0.0016191407
#> 13 U68497 U68498 0.12291500  0.0178006922 0.0011451050 0.0005798517
#> 14 U68497 U68499 0.11423944  0.0178657773 0.0010645577 0.0006536140
#> 15 U68497 U68500 0.14635158  0.0465803514 0.0014882273 0.0021421943
#> 16 U68497 U68501 0.15232262  0.0289560070 0.0015646067 0.0007113171
#> 17 U68497 U68502 0.18844835  0.1082484036 0.0019299056 0.0043143765
#> 18 U68497 U68503 0.15389053  0.0398280135 0.0015431390 0.0011502503
#> 19 U68497 U68504 0.15616528  0.1109611491 0.0015766080 0.0039369110
#> 20 U68497 U68505 0.15136945  0.0606006545 0.0015059746 0.0011657538
#> 21 U68497 U68506 0.16502115  0.0782677755 0.0016165774 0.0020628254
#> 22 U68497 U68507 0.20090792  0.0518729513 0.0020496504 0.0009519639
#> 23 U68497 U68508 0.20011790  0.0534260717 0.0021725951 0.0017341869
#> 24 U68498 U68499 0.07596391 -0.0004838245 0.0005620535 0.0003865916
#> 25 U68498 U68500 0.08840278  0.0463667477 0.0008260570 0.0016460465
#> 26 U68498 U68501 0.10286217  0.0101958272 0.0010525516 0.0003166930
#> 27 U68498 U68502 0.13009195  0.0940491643 0.0013057579 0.0044558026
#> 28 U68498 U68503 0.13641935  0.0441729344 0.0013004754 0.0016970196
#> 29 U68498 U68504 0.11024181  0.1184319156 0.0010976172 0.0049864168
#> 30 U68498 U68505 0.10652695  0.0712975708 0.0010391852 0.0023973803
#> 31 U68498 U68506 0.13085907  0.0870858286 0.0013133008 0.0029231671
#> 32 U68498 U68507 0.10809058  0.0543870183 0.0011142558 0.0015017030
#> 33 U68498 U68508 0.16297575  0.0440517806 0.0018854712 0.0010247788
#> 34 U68499 U68500 0.04753693  0.0102737435 0.0004383279 0.0005298658
#> 35 U68499 U68501 0.13602580  0.0097162829 0.0012647818 0.0007716744
#> 36 U68499 U68502 0.13681777  0.0901640402 0.0012227320 0.0042540042
#> 37 U68499 U68503 0.15100361  0.0458293630 0.0012941257 0.0020242775
#> 38 U68499 U68504 0.10903541  0.0941359554 0.0010512661 0.0039210531
#> 39 U68499 U68505 0.11155268  0.0408401022 0.0010437110 0.0009288677
#> 40 U68499 U68506 0.12931635  0.0641003604 0.0012583150 0.0020419709
#> 41 U68499 U68507 0.12466831  0.0567007476 0.0012199686 0.0017900222
#> 42 U68499 U68508 0.13610962  0.0898484362 0.0014962567 0.0041187582
#> 43 U68500 U68501 0.14964410  0.0582042332 0.0015776531 0.0020548121
#> 44 U68500 U68502 0.15988006  0.1438472932 0.0016078164 0.0064428544
#> 45 U68500 U68503 0.15673725  0.0861977448 0.0013892022 0.0037479479
#> 46 U68500 U68504 0.12800773  0.1677553950 0.0013867770 0.0080191008
#> 47 U68500 U68505 0.13063843  0.1085890328 0.0013768721 0.0039717993
#> 48 U68500 U68506 0.14653147  0.1236919447 0.0016022322 0.0051547115
#> 49 U68500 U68507 0.14494020  0.0963918335 0.0014435239 0.0034569489
#> 50 U68500 U68508 0.13591211  0.0806060567 0.0014251451 0.0030925806
#> 51 U68501 U68502 0.14933484  0.0672748223 0.0014052841 0.0034772435
#> 52 U68501 U68503 0.15573005  0.0554867582 0.0013851157 0.0021116389
#> 53 U68501 U68504 0.11320207  0.1420723073 0.0010409458 0.0060663691
#> 54 U68501 U68505 0.12691557  0.0670868311 0.0011168607 0.0020037645
#> 55 U68501 U68506 0.11939683  0.1092378646 0.0011341777 0.0038233059
#> 56 U68501 U68507 0.14111743  0.0661867231 0.0013055455 0.0018710869
#> 57 U68501 U68508 0.15062997  0.0674985077 0.0017283433 0.0012918646
#> 58 U68502 U68503 0.14127514  0.1679513190 0.0012163932 0.0074519537
#> 59 U68502 U68504 0.10809248  0.1522765380 0.0010257402 0.0058858088
#> 60 U68502 U68505 0.10525681  0.1405815063 0.0009871737 0.0056509830
#> 61 U68502 U68506 0.12796255  0.1874086342 0.0012356488 0.0084817971
#> 62 U68502 U68507 0.11030071  0.1651909774 0.0010889044 0.0071876798
#> 63 U68502 U68508 0.13517843  0.1841705167 0.0015314707 0.0079501818
#> 64 U68503 U68504 0.13509941  0.1269048145 0.0011113531 0.0047146155
#> 65 U68503 U68505 0.14189994  0.0424813552 0.0011379089 0.0009375791
#> 66 U68503 U68506 0.15642763  0.0962526692 0.0013415432 0.0027808415
#> 67 U68503 U68507 0.14632577  0.0425317911 0.0011497637 0.0008663924
#> 68 U68503 U68508 0.14702536  0.0786270338 0.0013787524 0.0025009601
#> 69 U68504 U68505 0.02018431  0.0935703588 0.0001353697 0.0042476907
#> 70 U68504 U68506 0.01963868  0.0616993490 0.0002675568 0.0019647032
#> 71 U68504 U68507 0.04645002  0.1121475416 0.0004670412 0.0041104637
#> 72 U68504 U68508 0.10262621  0.1452111714 0.0013299190 0.0066748916
#> 73 U68505 U68506 0.04029444  0.0848476525 0.0004104632 0.0024364085
#> 74 U68505 U68507 0.04627943  0.0203211276 0.0004322275 0.0003654116
#> 75 U68505 U68508 0.10915633  0.0474209893 0.0013291206 0.0021449341
#> 76 U68506 U68507 0.04958974  0.0826465317 0.0004228026 0.0022519366
#> 77 U68506 U68508 0.12249925  0.1139798033 0.0015342708 0.0043738539
#> 78 U68507 U68508 0.11267851  0.0468580415 0.0012934094 0.0019973539

dnastring2kaks(hiv, model = "NG86")

#>             Comp1 Comp2   seq1   seq2 Codons Compared Ambigiuous Indels Ns
#> result.1        1     2 U68496 U68497     91       91          0      0  0
#> result.2        1     3 U68496 U68498     91       91          0      0  0
#> result.3        1     4 U68496 U68499     91       91          0      0  0
#> result.4        1     5 U68496 U68500     91       91          0      0  0
#> result.5        1     6 U68496 U68501     91       91          0      0  0
#> result.6        1     7 U68496 U68502     91       91          0      0  0
#> result.7        1     8 U68496 U68503     91       91          0      0  0
#> result.8        1     9 U68496 U68504     91       91          0      0  0
#> result.9        1    10 U68496 U68505     91       91          0      0  0
#> result.10       1    11 U68496 U68506     91       91          0      0  0
#> result.11       1    12 U68496 U68507     91       91          0      0  0
#> result.12       1    13 U68496 U68508     91       91          0      0  0
#> result.1.1      2     3 U68497 U68498     91       91          0      0  0
#> result.2.1      2     4 U68497 U68499     91       91          0      0  0
#> result.3.1      2     5 U68497 U68500     91       91          0      0  0
#> result.4.1      2     6 U68497 U68501     91       91          0      0  0
#> result.5.1      2     7 U68497 U68502     91       91          0      0  0
#> result.6.1      2     8 U68497 U68503     91       91          0      0  0
#> result.7.1      2     9 U68497 U68504     91       91          0      0  0
#> result.8.1      2    10 U68497 U68505     91       91          0      0  0
#> result.9.1      2    11 U68497 U68506     91       91          0      0  0
#> result.10.1     2    12 U68497 U68507     91       91          0      0  0
#> result.11.1     2    13 U68497 U68508     91       91          0      0  0
#> result.1.2      3     4 U68498 U68499     91       91          0      0  0
#> result.2.2      3     5 U68498 U68500     91       91          0      0  0
#> result.3.2      3     6 U68498 U68501     91       91          0      0  0
#> result.4.2      3     7 U68498 U68502     91       91          0      0  0
#> result.5.2      3     8 U68498 U68503     91       91          0      0  0
#> result.6.2      3     9 U68498 U68504     91       91          0      0  0
#> result.7.2      3    10 U68498 U68505     91       91          0      0  0
#> result.8.2      3    11 U68498 U68506     91       91          0      0  0
#> result.9.2      3    12 U68498 U68507     91       91          0      0  0
#> result.10.2     3    13 U68498 U68508     91       91          0      0  0
#> result.1.3      4     5 U68499 U68500     91       91          0      0  0
#> result.2.3      4     6 U68499 U68501     91       91          0      0  0
#> result.3.3      4     7 U68499 U68502     91       91          0      0  0
#> result.4.3      4     8 U68499 U68503     91       91          0      0  0
#> result.5.3      4     9 U68499 U68504     91       91          0      0  0
#> result.6.3      4    10 U68499 U68505     91       91          0      0  0
#> result.7.3      4    11 U68499 U68506     91       91          0      0  0
#> result.8.3      4    12 U68499 U68507     91       91          0      0  0
#> result.9.3      4    13 U68499 U68508     91       91          0      0  0
#> result.1.4      5     6 U68500 U68501     91       91          0      0  0
#> result.2.4      5     7 U68500 U68502     91       91          0      0  0
#> result.3.4      5     8 U68500 U68503     91       91          0      0  0
#> result.4.4      5     9 U68500 U68504     91       91          0      0  0
#> result.5.4      5    10 U68500 U68505     91       91          0      0  0
#> result.6.4      5    11 U68500 U68506     91       91          0      0  0
#> result.7.4      5    12 U68500 U68507     91       91          0      0  0
#> result.8.4      5    13 U68500 U68508     91       91          0      0  0
#> result.1.5      6     7 U68501 U68502     91       91          0      0  0
#> result.2.5      6     8 U68501 U68503     91       91          0      0  0
#> result.3.5      6     9 U68501 U68504     91       91          0      0  0
#> result.4.5      6    10 U68501 U68505     91       91          0      0  0
#> result.5.5      6    11 U68501 U68506     91       91          0      0  0
#> result.6.5      6    12 U68501 U68507     91       91          0      0  0
#> result.7.5      6    13 U68501 U68508     91       91          0      0  0
#> result.1.6      7     8 U68502 U68503     91       91          0      0  0
#> result.2.6      7     9 U68502 U68504     91       91          0      0  0
#> result.3.6      7    10 U68502 U68505     91       91          0      0  0
#> result.4.6      7    11 U68502 U68506     91       91          0      0  0
#> result.5.6      7    12 U68502 U68507     91       91          0      0  0
#> result.6.6      7    13 U68502 U68508     91       91          0      0  0
#> result.1.7      8     9 U68503 U68504     91       91          0      0  0
#> result.2.7      8    10 U68503 U68505     91       91          0      0  0
#> result.3.7      8    11 U68503 U68506     91       91          0      0  0
#> result.4.7      8    12 U68503 U68507     91       91          0      0  0
#> result.5.7      8    13 U68503 U68508     91       91          0      0  0
#> result.1.8      9    10 U68504 U68505     91       91          0      0  0
#> result.2.8      9    11 U68504 U68506     91       91          0      0  0
#> result.3.8      9    12 U68504 U68507     91       91          0      0  0
#> result.4.8      9    13 U68504 U68508     91       91          0      0  0
#> result.1.9     10    11 U68505 U68506     91       91          0      0  0
#> result.2.9     10    12 U68505 U68507     91       91          0      0  0
#> result.3.9     10    13 U68505 U68508     91       91          0      0  0
#> result.1.10    11    12 U68506 U68507     91       91          0      0  0
#> result.2.10    11    13 U68506 U68508     91       91          0      0  0
#> result.12.1    12    13 U68507 U68508     91       91          0      0  0
#>                           Sd               Sn                S                N
#> result.1                   3                6               57              216
#> result.2                 1.5             18.5             57.5            215.5
#> result.3                 1.5             19.5             56.5            216.5
#> result.4                 2.5             25.5 56.1666666666667 216.833333333333
#> result.5                 2.5             23.5 57.3333333333333 215.666666666667
#> result.6    5.83333333333333 31.1666666666667             57.5            215.5
#> result.7                 3.5             28.5 57.1666666666667 215.833333333333
#> result.8                   4               26               57              216
#> result.9                   5               25 56.8333333333333 216.166666666667
#> result.10                2.5             26.5 57.1666666666667 215.833333333333
#> result.11                4.5             29.5 57.6666666666667 215.333333333333
#> result.12                4.5             36.5 57.1666666666667 215.833333333333
#> result.1.1               1.5             22.5             57.5            215.5
#> result.2.1               1.5             21.5             56.5            216.5
#> result.3.1               2.5             27.5 56.1666666666667 216.833333333333
#> result.4.1               2.5             27.5 57.3333333333333 215.666666666667
#> result.5.1  5.33333333333333 33.6666666666667             57.5            215.5
#> result.6.1               3.5             28.5 57.1666666666667 215.833333333333
#> result.7.1                 6               28               57              216
#> result.8.1                 5               27 56.8333333333333 216.166666666667
#> result.9.1               5.5             28.5 57.1666666666667 215.833333333333
#> result.10.1              4.5             33.5 57.6666666666667 215.333333333333
#> result.11.1              4.5             36.5 57.1666666666667 215.833333333333
#> result.1.2                 0               14               57              216
#> result.2.2               2.5             16.5 56.6666666666667 216.333333333333
#> result.3.2                 1               19 57.8333333333333 215.166666666667
#> result.4.2  3.83333333333333 24.1666666666667               58              215
#> result.5.2               2.5             24.5 57.6666666666667 215.333333333333
#> result.6.2                 5               20             57.5            215.5
#> result.7.2                 4               19 57.3333333333333 215.666666666667
#> result.8.2               4.5             23.5 57.6666666666667 215.333333333333
#> result.9.2               3.5             19.5 58.1666666666667 214.833333333333
#> result.10.2                4               30 57.6666666666667 215.333333333333
#> result.1.3                 1               10 55.6666666666667 217.333333333333
#> result.2.3                 1               25 56.8333333333333 216.166666666667
#> result.3.3  3.83333333333333 25.1666666666667               57              216
#> result.4.3               2.5             26.5 56.6666666666667 216.333333333333
#> result.5.3               4.5             20.5             56.5            216.5
#> result.6.3               3.5             20.5 56.3333333333333 216.666666666667
#> result.7.3                 4               24 56.6666666666667 216.333333333333
#> result.8.3               3.5             22.5 57.1666666666667 215.833333333333
#> result.9.3                 4               27 56.6666666666667 216.333333333333
#> result.1.4               3.5             27.5             56.5            216.5
#> result.2.4  6.33333333333333 29.6666666666667 56.6666666666667 216.333333333333
#> result.3.4                 4               28 56.3333333333333 216.666666666667
#> result.4.4               6.5             24.5 56.1666666666667 216.833333333333
#> result.5.4               5.5             24.5               56              217
#> result.6.4               5.5             27.5 56.3333333333333 216.666666666667
#> result.7.4                 5               26 56.8333333333333 216.166666666667
#> result.8.4                 5               26 56.3333333333333 216.666666666667
#> result.1.5  2.33333333333333 26.6666666666667 57.8333333333333 215.166666666667
#> result.2.5               3.5             27.5             57.5            215.5
#> result.3.5               6.5             20.5 57.3333333333333 215.666666666667
#> result.4.5               4.5             22.5 57.1666666666667 215.833333333333
#> result.5.5                 6               22             57.5            215.5
#> result.6.5               4.5             24.5               58              215
#> result.7.5                 6               28             57.5            215.5
#> result.1.6               7.5             25.5 57.6666666666667 215.333333333333
#> result.2.6  7.83333333333333 21.1666666666667             57.5            215.5
#> result.3.6  6.83333333333333 20.1666666666667 57.3333333333333 215.666666666667
#> result.4.6  8.33333333333333 24.6666666666667 57.6666666666667 215.333333333333
#> result.5.6  7.33333333333333 21.6666666666667 58.1666666666667 214.833333333333
#> result.6.6  8.33333333333333 26.6666666666667 57.6666666666667 215.333333333333
#> result.1.7                 7               23 57.1666666666667 215.833333333333
#> result.2.7                 4               24               57              216
#> result.3.7               6.5             26.5 57.3333333333333 215.666666666667
#> result.4.7                 4               24 57.8333333333333 215.166666666667
#> result.5.7                 5               27 57.3333333333333 215.666666666667
#> result.1.8                 4                4 56.8333333333333 216.166666666667
#> result.2.8               3.5              4.5 57.1666666666667 215.833333333333
#> result.3.8                 6                9 57.6666666666667 215.333333333333
#> result.4.8               6.5             21.5 57.1666666666667 215.833333333333
#> result.1.9               5.5              8.5               57              216
#> result.2.9                 2                9             57.5            215.5
#> result.3.9               2.5             22.5               57              216
#> result.1.10              5.5              9.5 57.8333333333333 215.166666666667
#> result.2.10                6               25 57.3333333333333 215.666666666667
#> result.12.1              2.5             22.5 57.8333333333333 215.166666666667
#>                             ps                 pn             pn/ps
#> result.1    0.0526315789473684 0.0277777777777778 0.527777777777778
#> result.2    0.0260869565217391   0.08584686774942   3.2907965970611
#> result.3    0.0265486725663717 0.0900692840646651  3.39260969976905
#> result.4    0.0445103857566766  0.117601844734819  2.64212144504228
#> result.5    0.0436046511627907  0.108964451313756  2.49891808346213
#> result.6     0.101449275362319  0.144624903325599  1.42558833278091
#> result.7    0.0612244897959184  0.132046332046332  2.15675675675676
#> result.8    0.0701754385964912   0.12037037037037  1.71527777777778
#> result.9    0.0879765395894428  0.115651503469545  1.31457208943716
#> result.10    0.043731778425656  0.122779922779923  2.80756756756757
#> result.11   0.0780346820809248  0.136996904024768  1.75558995528036
#> result.12   0.0787172011661808  0.169111969111969  2.14834834834835
#> result.1.1  0.0260869565217391  0.104408352668213  4.00232018561485
#> result.2.1  0.0265486725663717 0.0993071593533487  3.74056966897614
#> result.3.1  0.0445103857566766  0.126825518831668  2.84934665641814
#> result.4.1  0.0436046511627907  0.127511591962906   2.9242658423493
#> result.5.1  0.0927536231884058  0.156225831399845  1.68430974477958
#> result.6.1  0.0612244897959184  0.132046332046332  2.15675675675676
#> result.7.1   0.105263157894737   0.12962962962963  1.23148148148148
#> result.8.1  0.0879765395894428  0.124903623747109  1.41973785659214
#> result.9.1  0.0962099125364431  0.132046332046332  1.37248157248157
#> result.10.1 0.0780346820809248  0.155572755417957  1.99363605091159
#> result.11.1 0.0787172011661808  0.169111969111969  2.14834834834835
#> result.1.2                   0 0.0648148148148148              <NA>
#> result.2.2  0.0441176470588235  0.076271186440678  1.72881355932203
#> result.3.2  0.0172910662824208 0.0883036405886909  5.10689388071262
#> result.4.2  0.0660919540229885  0.112403100775194  1.70070778564206
#> result.5.2  0.0433526011560694  0.113777089783282  2.62445820433437
#> result.6.2  0.0869565217391304 0.0928074245939675  1.06728538283063
#> result.7.2  0.0697674418604651 0.0880989180834621  1.26275115919629
#> result.8.2  0.0780346820809248  0.109133126934985  1.39852081183351
#> result.9.2  0.0601719197707736 0.0907680372381691  1.50847833314862
#> result.10.2  0.069364161849711  0.139318885448916  2.00851393188855
#> result.1.3  0.0179640718562874 0.0460122699386503  2.56134969325153
#> result.2.3  0.0175953079178886  0.115651503469545  6.57286044718581
#> result.3.3  0.0672514619883041  0.116512345679012  1.73248792270531
#> result.4.3  0.0441176470588235  0.122496147919877  2.77657935285054
#> result.5.3   0.079646017699115 0.0946882217090069   1.1888632281242
#> result.6.3  0.0621301775147929 0.0946153846153846  1.52285714285714
#> result.7.3  0.0705882352941176  0.110939907550077  1.57164869029276
#> result.8.3  0.0612244897959184  0.104247104247104   1.7027027027027
#> result.9.3  0.0705882352941176  0.124807395993837  1.76810477657935
#> result.1.4  0.0619469026548673    0.1270207852194  2.05047838997031
#> result.2.4   0.111764705882353   0.13713405238829  1.22698888978996
#> result.3.4  0.0710059171597633  0.129230769230769              1.82
#> result.4.4   0.115727002967359  0.112990007686395 0.976349553597824
#> result.5.4  0.0982142857142857  0.112903225806452  1.14956011730205
#> result.6.4  0.0976331360946746  0.126923076923077               1.3
#> result.7.4  0.0879765395894428  0.120277563608327  1.36715497301465
#> result.8.4  0.0887573964497041               0.12             1.352
#> result.1.5  0.0403458213256484  0.123934934159566  3.07181586809782
#> result.2.5  0.0608695652173913  0.127610208816705  2.09645343056016
#> result.3.5   0.113372093023256 0.0950540958268934 0.838425870883367
#> result.4.5  0.0787172011661808  0.104247104247104  1.32432432432432
#> result.5.5   0.104347826086957  0.102088167053364 0.978344934261408
#> result.6.5  0.0775862068965517  0.113953488372093   1.4687338501292
#> result.7.5   0.104347826086957  0.129930394431555  1.24516627996906
#> result.1.6   0.130057803468208  0.118421052631579 0.910526315789474
#> result.2.6   0.136231884057971 0.0982211910286156  0.72098533840154
#> result.3.6   0.119186046511628 0.0935085007727975 0.784559128435179
#> result.4.6   0.144508670520231  0.114551083591331 0.792693498452012
#> result.5.6   0.126074498567335  0.100853374709077 0.799950631215178
#> result.6.6   0.144508670520231  0.123839009287926 0.856965944272446
#> result.1.7   0.122448979591837  0.106563706563707  0.87027027027027
#> result.2.7  0.0701754385964912  0.111111111111111  1.58333333333333
#> result.3.7   0.113372093023256  0.122874806800618  1.08381880870289
#> result.4.7   0.069164265129683   0.11154144074361  1.61270333075136
#> result.5.7  0.0872093023255814  0.125193199381762   1.4355486862442
#> result.1.8  0.0703812316715543 0.0185042405551272 0.262914417887433
#> result.2.8  0.0612244897959184 0.0208494208494209 0.340540540540541
#> result.3.8   0.104046242774566 0.0417956656346749 0.401702786377709
#> result.4.8   0.113702623906706 0.0996138996138996 0.876091476091476
#> result.1.9  0.0964912280701754 0.0393518518518519 0.407828282828283
#> result.2.9  0.0347826086956522 0.0417633410672854  1.20069605568445
#> result.3.9   0.043859649122807  0.104166666666667             2.375
#> result.1.10 0.0951008645533141 0.0441518202943455 0.464263080064784
#> result.2.10  0.104651162790698  0.115919629057187  1.10767645543534
#> result.12.1 0.0432276657060519  0.104570100697134  2.41905499612703
#>                             ds                 dn             dn/ds
#> result.1    0.0545695157118212 0.0283052459871352 0.518700699793888
#> result.2     0.026551445288187 0.0911703470876381  3.43372445823884
#> result.3    0.0270299523623977 0.0959537646568416  3.54990505977829
#> result.4    0.0458858673563109  0.127915513677956  2.78768869474935
#> result.5    0.0449236061858017  0.117741219733924  2.62092093067847
#> result.6     0.108999740568461  0.160668709585171  1.47402836692312
#> result.7    0.0638683562552301  0.145244792123914  2.27412760622002
#> result.8    0.0736788296875599  0.131206087122249   1.7807840824649
#> result.9    0.0935791594009808   0.12560604303091  1.34224376276662
#> result.10   0.0450585929574625  0.134081795349486  2.97572086807206
#> result.11   0.0823998581838992  0.151277415057255   1.8358941191335
#> result.12   0.0831620246840571  0.191636389628777  2.30437378547273
#> result.1.1   0.026551445288187  0.112429520622234  4.23440303915417
#> result.2.1  0.0270299523623977  0.106526626973117  3.94105862803185
#> result.3.1  0.0458858673563109  0.138934995654973  3.02783849711548
#> result.4.1  0.0449236061858017  0.139761149994027   3.1110848362436
#> result.5.1  0.0990106916882497  0.175180610249359  1.76931003371779
#> result.6.1  0.0638683562552301  0.145244792123914  2.27412760622002
#> result.7.1   0.113423227292943  0.142317401461104  1.25474653523599
#> result.8.1  0.0935791594009808  0.136625525007927  1.45999949008406
#> result.9.1    0.10296515533144  0.145244792123914   1.4106208227082
#> result.10.1 0.0823998581838992  0.174356159057321  2.11597644583555
#> result.11.1 0.0831620246840571  0.191636389628777  2.30437378547273
#> result.1.2                   0 0.0677880461012017              <NA>
#> result.2.2  0.0454684663623261 0.0804341477651982  1.76900947404383
#> result.3.2  0.0174935044248067 0.0939498205495953  5.37055459375937
#> result.4.2   0.069187300238414  0.121775206698307  1.76008033669011
#> result.5.2  0.0446560452682624   0.12339316262594  2.76319055762057
#> result.6.2  0.0924244803179611 0.0990720883078883  1.07192475377798
#> result.7.2   0.073228852172937 0.0937178136336677  1.27979356295718
#> result.8.2  0.0823998581838992   0.11793859312913   1.4312960692956
#> result.9.2  0.0627230993043084 0.0967483064333858  1.54246693014961
#> result.10.2 0.0727843418621031  0.154123718283669  2.11753949188235
#> result.1.3  0.0181827087074946 0.0474842095607451   2.6115036172346
#> result.2.3  0.0178049904773047   0.12560604303091  7.05454143269636
#> result.3.3  0.0704599406264042  0.126624521256008  1.79711365252779
#> result.4.3  0.0454684663623261  0.133742547577616  2.94143520284711
#> result.5.3  0.0842004764033108  0.101221565066322  1.20214955294888
#> result.6.3  0.0648551985028701  0.101138208126399  1.55944643546073
#> result.7.3  0.0741343759774745  0.120056036386316  1.61943814597961
#> result.8.3  0.0638683562552301  0.112242217679564  1.75739950517942
#> result.9.3  0.0741343759774745  0.136510078393206  1.84138702988051
#> result.1.4  0.0646553964009121  0.139170038559349  2.15248913944305
#> result.2.4   0.121024643713475   0.15144523277581  1.25135863348919
#> result.3.4  0.0745955951479785  0.141835352045924  1.90139044758017
#> result.4.4   0.125695312562783  0.122465898420489  0.97430760084485
#> result.5.4   0.105268018271192  0.122363730514676  1.16240176764268
#> result.6.4   0.104599596108253  0.139052417498522  1.32937814936315
#> result.7.4  0.0935791594009808  0.131095546037613  1.40090536051811
#> result.8.4  0.0944643067332107  0.130765040358583  1.38427989238194
#> result.1.5    0.04147157058216  0.135464176341989   3.2664346789958
#> result.2.5  0.0634819823655594  0.139879977105869  2.20345949974237
#> result.3.5   0.122915891328837  0.101640422515845 0.826910348344833
#> result.4.5  0.0831620246840571  0.112242217679564  1.34968115682592
#> result.5.5   0.112359208671495  0.109738934776993 0.976679491378732
#> result.6.5   0.081899468973744  0.123601135780343  1.50918116233414
#> result.7.5   0.112359208671495  0.142681100810109  1.26986566118729
#> result.1.6   0.142835223178404  0.128887692694994 0.902352303773218
#> result.2.6   0.150342008285561   0.10527596415443 0.700243167927274
#> result.3.6   0.129796673535868 0.0998725961228087 0.769454204041751
#> result.4.6   0.160524722432813  0.124306126828754 0.774373722283066
#> result.5.6   0.138031675513526  0.108310942796978 0.784681794189803
#> result.6.6   0.160524722432813  0.135349271124206 0.843167763027005
#> result.1.7   0.133686173554739  0.114937638370506 0.859757111108007
#> result.2.7  0.0736788296875599  0.120256987556385  1.63217830774922
#> result.3.7   0.122915891328837  0.134195261740814  1.09176494829136
#> result.4.7  0.0725641058409195  0.120762327332266  1.66421574320794
#> result.5.7  0.0927104669753825  0.136973042693928  1.47742803118766
#> result.1.8  0.0739059002893285 0.0187363373661175 0.253516123784002
#> result.2.8  0.0638683562552301 0.0211447050669682 0.331066999477331
#> result.3.8    0.11200896641424 0.0430054099614337 0.383946136976105
#> result.4.8   0.123305384995836  0.106880264546396 0.866793161953194
#> result.1.9   0.103287937917901 0.0404218261538492 0.391350887322186
#> result.2.9  0.0356149996994909 0.0429711784958646   1.2065472092782
#> result.3.9  0.0451943936386476  0.112148800478223  2.48147594090785
#> result.1.10  0.101693980828118 0.0455047763039953 0.447467745223851
#> result.2.10  0.112711652287003   0.12592311905919  1.11721473782094
#> result.12.1 0.0445234567872044  0.112617450905259  2.52939594163822

if (FALSE) {
dnastring2kaks(hiv, model = "NG86", threads = 2)
}
if (FALSE) {
dnastring2kaks(hiv, model = "NG86", align = TRUE, threads = 2)
dnastring2kaks(hiv, model = "Li", align = TRUE, substitutionMatrix = "BLOSUM45")
}